pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==24.1.0",
    "aiogram==3.16.0",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "apscheduler==3.11.0",
    "asyncpg==0.30.0",
    "attrs==24.3.0",
    "certifi==2024.12.14",
    "frozenlist==1.5.0",
    "idna==3.10",
    "magic-filter==1.0.12",
    "multidict==6.1.0",
    "propcache==0.2.1",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "python-decouple>=3.8",
    "python-slugify>=8.0.4",
    "sqlalchemy==2.0.36",
    "typing-extensions==4.12.2",
    "tzlocal==5.2",
    "yarl==1.18.3",
]
